Change Accessor design: pass database on construct
This commit... was a little sad. I realized I had a design flaw, and to
fix it required me to change the design of the Accessors. Now whenever
Accessors are used, they need to be given a database to work on, and are
(in an abstract sense) just database shortcuts for certain data types.

Noteworthy changes:
- Accessors can no longer be used in try-with-resources blocks
(since the database is created outside)
- Accessors require a Database instance on construction
- User/Person Accessor tests have been updated with the above changes
in mind